+++ 2.5-D Game/Worlds/TileAtlasLoader.cpp
@@ -38,12 +38,12 @@ bool Game::TileAtlasLoader::loadAtlas(std::string path, TileAtlas& atlas) {
 		while (std::getline(in, line)) {
 			while (line.length() > 0) {
 				char character = line.at(0);
-				unsigned long long charactersToErase = 1;
+				size_t charactersToErase = 1;
 
 				if (character == '<') {
 					currentTag = "";
 
-					unsigned long long j = line.find(' ');
+					size_t j = line.find(' ');
 					if (j != std::string::npos) {
 						currentTag += line.substr(1, j - 1);
 					}
@@ -101,13 +101,13 @@ bool Game::TileAtlasLoader::loadAtlas(std::string path, TileAtlas& atlas) {
 				} else if (character != ' ' && currentTag != "") {
 					std::string propertyType = "";
 
-					unsigned long long j = line.find('=');
+					size_t j = line.find('=');
 					if (j != std::string::npos) {
 						propertyType += line.substr(0, j);
 					}
 
 					j = line.find('"');
-					unsigned long long k = line.find('"', j + 1);
+					size_t k = line.find('"', j + 1);
 
 					if (j != std::string::npos && k != std::string::npos) {
 						if ((propertyType == "material" || propertyType == "Material") && (currentTag == "Tile" || currentTag == "tile")) {
